backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
gunicorn==23.0.0
Werkzeug==3.0.3
